:recycle: refactor: directory (related: #5)
