+++ src/main/java/com/server/domain/user/repository/UserRepository.java
@@ -1,9 +1,9 @@
-package com.server.repository;
+package com.server.domain.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.server.model.User;
+import com.server.domain.user.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, String> {
