--- src/main/resources/application-test.properties
@@ -1,8 +1,0 @@
-spring.application.name=co-co-gong-server
-
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.datasource.driverClassName=org.h2.Driver
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
